Bug 568393 - msvc 2008 compiler generates bad code for CallInfo static initializer (r+wmaddox)
ugh

UltraBlame original commit: 3b300e72d88bfcd101673ebaca6dd0afdf10c0de

+++ js/src/nanojit/LIR.h
@@ -323,7 +323,7 @@ namespace nanojit
         uintptr_t   _address;
         uint32_t    _typesig:27;     
         AbiKind     _abi:3;
-        uint8_t     _isPure:1;      
+        uint32_t    _isPure:1;      
         AccSet      _storeAccSet;   
         verbose_only ( const char* _name; )
